+++ R/reexports.R
@@ -1,0 +1,3 @@
+#' @importFrom stats density
+#' @export
+stats::density