+++ NEWS.md
@@ -9,6 +9,7 @@
   * `cal_age_range()`, `cal_age_min()`, and `cal_age_max()` for simple ranges
 * Added aggregation functions for for calibrated dates (`cal` class):
   * `cal_sum()` for summed probability
+  * `cal_density()` for composite kernel density estimation
 * Added classes, functions and datasets for calibration curves:
   * `c14_curve` is a vctrs record class representing a calibration curve, with
     subclasses `c14_curve_14c` and `c14_curve_f14c` for CRA-based and F14C-based